+++ src/main/java/ru/itmentor/spring/boot_security/demo/service/UserService.java
@@ -11,9 +11,9 @@ import java.util.Set;
 public interface UserService {
     List<User> getAllUsers();
     void add(User user);
-    User getUserById(int id);
+    User getUserById(long id);
     void updateUser(User updateUser);
-    void deleteUser(int id);
+    void deleteUser(long id);
     Set<Role> getSetOfRoles(List<String> id);
     User findByUserName(String name);
     User getUserByEmail(String email);
